New scrolling background script and improved movement speed to enchance movement on both directions (#41)
Background now scrolls onto a quad in both directions, initially the ship is in the middle, when it moves to a side it flips and the positioning on the x axis moves to the edge so it simulates acceleration. When there is no input anymore the ship returns back to the middle.