Coding repo (#43)
* New scrolling background script and improved movement speed to enchance movement on both directions

Background now scrolls onto a quad in both directions, initially the ship is in the middle, when it moves to a side it flips and the positioning on the x axis moves to the edge so it simulates acceleration. When there is no input anymore the ship returns back to the middle.

* Shooting and heat bar UI mechanic implemented

The ship shoots with a 0.5s cooldown and the heat bar goes down over time. When the bar is empty the player can't shoot anymore and needs to wait for the bar to fill but not necessarely fully.

* Bug fixing

There was a weird delay between the pressing button down and the way the shooting and heatbar mechanic occurs. Seems fixed for the most part

* Merge conflict fix attempt